requirements.txt
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==1.15.0
dnspython3==1.15.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mcstatus==4.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
